move non-opa related test code to utils class
