+++ src/main/java/tech/stackable/hbase/OpenPolicyAgentAccessController.java
@@ -1,11 +1,21 @@
 package tech.stackable.hbase;
 
+import com.fasterxml.jackson.annotation.JsonAutoDetect;
+import com.fasterxml.jackson.annotation.PropertyAccessor;
+import com.fasterxml.jackson.databind.DeserializationFeature;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.SerializationFeature;
+import com.google.protobuf.RpcCallback;
+import com.google.protobuf.RpcController;
 import java.io.IOException;
+import java.net.URI;
+import java.net.http.HttpClient;
 import java.util.*;
 import org.apache.hadoop.hbase.*;
 import org.apache.hadoop.hbase.client.*;
 import org.apache.hadoop.hbase.coprocessor.*;
 import org.apache.hadoop.hbase.io.hfile.HFile;
+import org.apache.hadoop.hbase.protobuf.generated.AccessControlProtos;
 import org.apache.hadoop.hbase.security.User;
 import org.apache.hadoop.hbase.security.UserProvider;
 import org.apache.hadoop.hbase.security.access.*;
@@ -13,17 +23,40 @@ import org.apache.hadoop.hbase.wal.WALEdit;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class OpenPolicyAgentAccessController extends AccessController {
+public class OpenPolicyAgentAccessController
+    implements MasterCoprocessor,
+        RegionCoprocessor,
+        RegionServerCoprocessor,
+        AccessControlProtos.AccessControlService.Interface,
+        MasterObserver,
+        RegionObserver,
+        RegionServerObserver,
+        EndpointObserver,
+        BulkLoadObserver {
   private static final Logger LOG = LoggerFactory.getLogger(OpenPolicyAgentAccessController.class);
 
   private UserProvider userProvider;
   private boolean authorizationEnabled;
   private boolean cellFeaturesEnabled;
 
+  // Opa-related
+  public static final String OPA_POLICY_URL_PROP = "hbase.security.authorization.opa.policy.url";
+  private final HttpClient httpClient = HttpClient.newHttpClient();
+  private URI opaUri;
+  private final ObjectMapper json;
+
+  public OpenPolicyAgentAccessController() {
+    this.json =
+        new ObjectMapper()
+            .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
+            .configure(SerializationFeature.FAIL_ON_EMPTY_BEANS, false)
+            .setVisibility(PropertyAccessor.ALL, JsonAutoDetect.Visibility.NONE)
+            .setVisibility(PropertyAccessor.FIELD, JsonAutoDetect.Visibility.PUBLIC_ONLY)
+            .setVisibility(PropertyAccessor.GETTER, JsonAutoDetect.Visibility.PUBLIC_ONLY);
+  }
+
   @Override
   public void start(CoprocessorEnvironment env) throws IOException {
-    super.start(env);
-
     authorizationEnabled = AccessChecker.isAuthorizationSupported(env.getConfiguration());
     if (!authorizationEnabled) {
       LOG.warn(
@@ -34,14 +67,27 @@ public class OpenPolicyAgentAccessController extends AccessController {
         (HFile.getFormatVersion(env.getConfiguration()) >= HFile.MIN_FORMAT_VERSION_WITH_TAGS);
     if (!cellFeaturesEnabled) {
       LOG.info(
-          "A minimum HFile version of "
-              + HFile.MIN_FORMAT_VERSION_WITH_TAGS
-              + " is required to persist cell ACLs. Consider setting "
-              + HFile.FORMAT_VERSION_KEY
-              + " accordingly.");
+          "A minimum HFile version of [{}] is required to persist cell ACLs. "
+              + "Consider setting [{}] accordingly.",
+          HFile.MIN_FORMAT_VERSION_WITH_TAGS,
+          HFile.FORMAT_VERSION_KEY);
     }
     // set the user-provider.
     this.userProvider = UserProvider.instantiate(env.getConfiguration());
+
+    // opa-related
+    if (authorizationEnabled) {
+      String opaPolicyUrl = env.getConfiguration().get(OPA_POLICY_URL_PROP);
+      if (opaPolicyUrl == null) {
+        throw new OpaException.UriMissing(OPA_POLICY_URL_PROP);
+      }
+
+      try {
+        this.opaUri = URI.create(opaPolicyUrl);
+      } catch (Exception e) {
+        throw new OpaException.UriInvalid(opaUri, e);
+      }
+    }
   }
 
   // TODO replace with user that returns the whole name for getShortName()
@@ -90,5 +136,61 @@ public class OpenPolicyAgentAccessController extends AccessController {
     User user = getActiveUser(c);
     LOG.info("preAppend: start with [{}]", user);
     return null;
+  }
+
+  @Override
+  public void grant(
+      RpcController controller,
+      AccessControlProtos.GrantRequest request,
+      RpcCallback<AccessControlProtos.GrantResponse> done) {}
+
+  @Override
+  public void revoke(
+      RpcController controller,
+      AccessControlProtos.RevokeRequest request,
+      RpcCallback<AccessControlProtos.RevokeResponse> done) {}
+
+  @Override
+  public void getUserPermissions(
+      RpcController controller,
+      AccessControlProtos.GetUserPermissionsRequest request,
+      RpcCallback<AccessControlProtos.GetUserPermissionsResponse> done) {}
+
+  @Override
+  public void checkPermissions(
+      RpcController controller,
+      AccessControlProtos.CheckPermissionsRequest request,
+      RpcCallback<AccessControlProtos.CheckPermissionsResponse> done) {}
+
+  @Override
+  public void hasPermission(
+      RpcController controller,
+      AccessControlProtos.HasPermissionRequest request,
+      RpcCallback<AccessControlProtos.HasPermissionResponse> done) {}
+
+  /*********************************** Observer/Service Getters ***********************************/
+  @Override
+  public Optional<RegionObserver> getRegionObserver() {
+    return Optional.of(this);
+  }
+
+  @Override
+  public Optional<MasterObserver> getMasterObserver() {
+    return Optional.of(this);
+  }
+
+  @Override
+  public Optional<EndpointObserver> getEndpointObserver() {
+    return Optional.of(this);
+  }
+
+  @Override
+  public Optional<BulkLoadObserver> getBulkLoadObserver() {
+    return Optional.of(this);
+  }
+
+  @Override
+  public Optional<RegionServerObserver> getRegionServerObserver() {
+    return Optional.of(this);
   }
 }
